# specifications/openapi/txlio-agentic-v0.yaml
openapi: 3.1.0
info:
  title: Txlio Agentic API
  version: 0.1.0
  description: AI-facing contract ingestion and timeline workflows.
servers:
  - url: http://localhost:3000
paths:
...
  /api/ai/tools/execute:
    post:
      operationId: executeAiTool
      summary: Execute a catalog tool against timeline APIs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - tool
                - input
              properties:
                tool:
                  type: string
                  enum:
                    - upload_contract
                    - get_timeline
                    - export_timeline_pdf
                input:
                  type: object
                  additionalProperties: true
      responses:
        '200':
          description: Tool execution result
        '400':
          description: Invalid tool input
        '401':
          description: Unauthorized
        '500':
          description: Tool execution failure